+++ CHANGELOG.md
@@ -13,6 +13,9 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-01-XX
+  * Added finite clines considerations in the stretching tool
+  * Added definition behavior fields in outliner form
 ### 2026-01-04
   * Improved Smart Tools startup
 ### 2026-01-04